==> 02-vf-sdk-examples/03-vertex-forecast-train-sdk.ipynb <==
# Vertex Forecast - training with SDK

import os

GCP_PROJECTS = !gcloud config get-value project
PROJECT_ID = GCP_PROJECTS[0]
PROJECT_NUM = !gcloud projects list --filter="$PROJECT_ID" --format="value(PROJECT_NUMBER)"
PROJECT_NUM = PROJECT_NUM[0]
LOCATION = 'us-central1'
BQ_LOCATION='US'

# TODO: Service Account address
VERTEX_SA = '934903580331-compute@developer.gserviceaccount.com' 

print(f"PROJECT_ID: {PROJECT_ID}")
print(f"PROJECT_NUM: {PROJECT_NUM}")
print(f"LOCATION: {LOCATION}")

import google.cloud.aiplatform as vertex_ai
from google.cloud import bigquery
from google.cloud import storage

import matplotlib.pyplot as plt
import pandas as pd
from datetime import datetime, timedelta

bq_client = bigquery.Client(
    project=PROJECT_ID, 
    location=BQ_LOCATION
)

storage_client = storage.Client(project=PROJECT_ID)

vertex_ai.init(
    project=PROJECT_ID,
    location=LOCATION
)

# previously defined
BQ_DATASET="m5_us"
BQ_TABLE="sdk_train"
BQ_TABLE_PLAN="sdk_plan"

# new vars
EXPERIMENT_tag="nb3"
VERSION='v1'

EXPERIMENT_NAME = f"m5_{EXPERIMENT_tag}_{VERSION}"
print(f'EXPERIMENT_NAME: {EXPERIMENT_NAME}')

## Create Vertex Managed Dataset 
* link to BigQuery Table

Reference for [`aiplatform.TimeSeriesDataset.create()`](https://googleapis.dev/python/aiplatform/latest/aiplatform.html#google.cloud.aiplatform.TimeSeriesDataset.create)

dataset = vertex_ai.TimeSeriesDataset.create(
    display_name = f'{EXPERIMENT_NAME}', 
    bq_source = f'bq://{PROJECT_ID}.{BQ_DATASET}.{BQ_TABLE}_prepped',
    labels = {'experiment':f'{EXPERIMENT_NAME}'}
)

dataset = vertex_ai.TimeSeriesDataset('projects/934903580331/locations/us-central1/datasets/462153324456574976')

# Train Forecasting Model with AutoML

Reference for [`aiplatform.AutoMLForecastingTrainingJob`](https://googleapis.dev/python/aiplatform/latest/aiplatform.html#google.cloud.aiplatform.AutoMLForecastingTrainingJob)

## train job config

### column specs

dataset.column_names

### define target, time, and ID

TARGET_COLUMN = 'gross_quantity'
TIME_COLUMN = 'date'
SERIES_COLUMN = 'timeseries_id'
PREDEFINED_SPLIT_COL = 'splits'

### features which will be available at forecast (inference)

# column_specs = list(set(dataset.column_names) - set(['splits','timeseries_id']))
# column_specs = dict.fromkeys(column_specs, 'auto')

AVAILABLE_AT_FORECAST_COLS = list(set(dataset.column_names) - set(['splits','timeseries_id','gross_quantity']))
AVAILABLE_AT_FORECAST_COLS

### features which will be unavailable at forecast (inference)

UNAVAILABLE_AT_FORECAST_COLS=[
    # TARGET_COLUMN,
    # 'sell_price'
]

ALL = [
    PREDEFINED_SPLIT_COL,
    # SERIES_COLUMN,
    # TARGET_COLUMN,
    TIME_COLUMN
]
# for col in UNAVAILABLE_AT_FORECAST_COLS:
#     ALL.append(col)
    
for col in AVAILABLE_AT_FORECAST_COLS:
    ALL.append(col)
ALL

Create a dictionary containing all features

COL_TRANSFORMS = {
    TIME_COLUMN:"timestamp",
    TARGET_COLUMN:"numeric",
    "product_id":"categorical",
    "location_id":"categorical",
    "weekday":"categorical",
    "event_name_1":"categorical",
    "year":"categorical",
    "event_type_1":"categorical",
    "month":"categorical",
    "dept_id":"categorical",
    "event_type_2":"categorical",
    "wday":"categorical",
    "state_id":"categorical",
    "snap_WI":"categorical",
    "snap_CA":"categorical",
    "snap_TX":"categorical",
    "event_name_2":"categorical",
    "cat_id":"categorical",
    "sell_price":"numeric"
}

import pickle as pkl
from pprint import pprint

LOCAL_COL_FILE = 'column_specs.pkl'

column_specs_dict = {
    "column_specs": COL_TRANSFORMS,
    "unavailable_at_forecast_columns": UNAVAILABLE_AT_FORECAST_COLS,
    "available_at_forecast_columns": AVAILABLE_AT_FORECAST_COLS,
    "time_series_identifier_column": SERIES_COLUMN,
    "time_column": TIME_COLUMN,
    "target_column": TARGET_COLUMN,
    "predefined_split_column_name": PREDEFINED_SPLIT_COL,
    # "time_series_attribute_columns": XXX
}

filehandler = open(LOCAL_COL_FILE, 'wb')
pkl.dump(column_specs_dict, filehandler)
filehandler.close()

# sanity check
# pprint(column_specs_dict)

### model config

**optimization_objective:**
* "minimize-rmse" (default) - Minimize root-mean-squared error (RMSE).
* "minimize-mae" - Minimize mean-absolute error (MAE).
* "minimize-rmsle" - Minimize root-mean-squared log error (RMSLE).
* "minimize-rmspe" - Minimize root-mean-squared percentage error (RMSPE).
* "minimize-wape-mae" - Minimize the combination of weighted absolute percentage error (WAPE) and mean-absolute-error (MAE).
* "minimize-quantile-loss" - Minimize the quantile loss at the defined quantiles. (Set this objective to build quantile forecasts.)

# TODO - edit these

# forecast spec
FORECAST_GRANULARITY = 'DAY'
DATA_GRANULARITY_COUNT=1
FORECAST_HORIZON = 14
CONTEXT_WINDOW = 14
forecast_test_length = 14
forecast_val_length = 14

# model config
OPTIMIZATION_OBJECTIVE="minimize-rmse"

# job spec
MILLI_NODE_HRS=1000
HOLIDAY_REGIONS=['GLOBAL', 'NA', 'US']

# export eval set BQ destination
BQ_EVAL_DESTINATION = f"bq://{PROJECT_ID}:{BQ_DATASET}:{BQ_TABLE}_automl_{VERSION}",

print(f"EXPERIMENT_NAME: {EXPERIMENT_NAME}")
print(f"VERSION: {VERSION}")
print(f"OPTIMIZATION_OBJECTIVE: {OPTIMIZATION_OBJECTIVE}")
print(f"TARGET_COLUMN: {TARGET_COLUMN}")
print(f"TIME_COLUMN: {TIME_COLUMN}")
print(f"SERIES_COLUMN: {SERIES_COLUMN}")
print(f"AVAILABLE_AT_FORECAST_COLS: {AVAILABLE_AT_FORECAST_COLS}")
print(f"FORECAST_HORIZON: {FORECAST_HORIZON}")
print(f"FORECAST_GRANULARITY: {FORECAST_GRANULARITY.lower()}")
print(f"CONTEXT_WINDOW: {CONTEXT_WINDOW}")
print(f"TARGET_COLUMN: {TARGET_COLUMN}")
print(f"BQ_EVAL_DESTINATION: {BQ_EVAL_DESTINATION[0]}")

UNAVAILABLE_AT_FORECAST_COLS='gross_quantity'

## create and submit job

forecast_job = vertex_ai.AutoMLForecastingTrainingJob(
    display_name = f'{EXPERIMENT_NAME}_training',
    optimization_objective=OPTIMIZATION_OBJECTIVE,
    column_specs = COL_TRANSFORMS,
    labels = {'experiment':f'{EXPERIMENT_NAME}'}
)

# DEST_JT=BQ_EVAL_DESTINATION[0]
# DEST_JT
BQ_EVAL_DESTINATION[0]

forecast=forecast_job.run(
    dataset=dataset,
    target_column=TARGET_COLUMN,
    time_column=TIME_COLUMN,
    time_series_identifier_column=SERIES_COLUMN,
    unavailable_at_forecast_columns=UNAVAILABLE_AT_FORECAST_COLS,
    available_at_forecast_columns=AVAILABLE_AT_FORECAST_COLS,
    forecast_horizon=FORECAST_HORIZON,
    data_granularity_unit=FORECAST_GRANULARITY.lower(),
    data_granularity_count=DATA_GRANULARITY_COUNT,
    predefined_split_column_name="splits",
    context_window = CONTEXT_WINDOW,
    export_evaluated_data_items=True,
    export_evaluated_data_items_bigquery_destination_uri='bq://hybrid-vertex.m5_us.sdk_train_automl_v1', #BQ_EVAL_DESTINATION[0],
    validation_options="fail-pipeline",
    budget_milli_node_hours = MILLI_NODE_HRS,
    model_display_name=f"{EXPERIMENT_NAME}_{BQ_TABLE}",
    model_labels={'experiment':f'{EXPERIMENT_NAME}'},
    holiday_regions=HOLIDAY_REGIONS,
    sync=True
)

FORECAST_MODEL_RSC_NAME = forecast.resource_name
print(f"FORECAST_MODEL_RSC_NAME: {FORECAST_MODEL_RSC_NAME}")

## default Model Evaluation

forecast_EVALS = forecast.list_model_evaluations()

for model_evaluation in forecast_EVALS:
    print(model_evaluation.to_dict())

# # Get evaluations
# model_evaluations = forecast.list_model_evaluations()

# # Print the evaluation metrics
# for evaluation in model_evaluations:
#     evaluation = evaluation.to_dict()
#     print("Model's evaluation metrics from training:\n")
#     metrics = evaluation["metrics"]
#     for metric in metrics.keys():
#         print(f"metric: {metric}, value: {metrics[metric]}\n")

### retrieve default model evaluation metrics

model_evaluation = list(forecast.list_model_evaluations())[0]
metrics_dict = {k: [v] for k, v in dict(model_evaluation.metrics).items()}
metrics_dict

## log metrics to Vertex AI Experiments

from datetime import datetime

# create run name
TIMESTAMP = datetime.now().strftime("%Y%m%d%H%M%S")
EXPERIMENT_RUN_NAME = f"run-{TIMESTAMP}"

# log params and metrics to dicts
params = {}
params["budget_hrs"] = MILLI_NODE_HRS
params["horizon"] = FORECAST_HORIZON
params["context_window"] = CONTEXT_WINDOW

metrics = {}
metrics["MAE"] = metrics_dict['meanAbsoluteError'][0]
metrics["RMSE"] = metrics_dict['rootMeanSquaredError'][0]
metrics["MAPE"] = metrics_dict['meanAbsolutePercentageError'][0]
metrics["rSquared"] = metrics_dict['rSquared'][0]
metrics["RMSLE"] = metrics_dict['rootMeanSquaredLogError'][0]
metrics["WAPE"] = metrics_dict['weightedAbsolutePercentageError'][0]

# # Create and log experiment
vertex_ai.init(experiment=EXPERIMENT_NAME.replace("_","-"))

with vertex_ai.start_run(EXPERIMENT_RUN_NAME) as my_run:
    my_run.log_metrics(metrics)
    my_run.log_params(params)

    vertex_ai.end_run()

## Using The Results

query = f"""
WITH
    RAW AS (
        SELECT 
            DATE({TIME_COLUMN}) as {TIME_COLUMN}, 
            DATE(predicted_on_date) as predicted_on_date, 
            CAST({TARGET_COLUMN} as INT64) AS {TARGET_COLUMN}, 
            #splits, 
            {SERIES_COLUMN}, 
            predicted_{TARGET_COLUMN}.value as predicted_{TARGET_COLUMN}
        FROM `{PROJECT_ID}.{BQ_DATASET}.{BQ_TABLE}_automl`
    ),
    LEAD AS (
        SELECT 
            *, 
            DATE_DIFF({TIME_COLUMN}, predicted_on_date, {FORECAST_GRANULARITY}) as prediction_lead_days
        FROM RAW
    ),
    LEFTSIDE AS (
        SELECT 
            {SERIES_COLUMN}, 
            {TIME_COLUMN}, 
            min(prediction_lead_days) as prediction_lead_days
        FROM LEAD
        GROUP BY {SERIES_COLUMN}, {TIME_COLUMN}
    )
SELECT *
FROM LEFTSIDE
LEFT OUTER JOIN LEAD
USING ({SERIES_COLUMN}, {TIME_COLUMN}, prediction_lead_days)
"""
autoML = bq_client.query(query).to_dataframe()
autoML

# print to run in BigQuery console
# print(query)

### Review Custom Metrics with SQL

Some common metrics for evaluating forecasting effectiveness are 
- MAPE, or Mean Absolute Percentage Error
    - $\textrm{MAPE} = \frac{1}{n}\sum{\frac{\mid(actual - forecast)\mid}{actual}}$
- MAE, or Mean Absolute Error
     - $\textrm{MAE} = \frac{1}{n}\sum{\mid(actual - forecast)\mid}$
- MAE divided by average demand so it yields a % like MAPE
    - $\textrm{pMAE} = \frac{\sum{\mid(actual - forecast)\mid}}{\sum{actual}}$
- MSE, or Mean Squared Error
    - $\textrm{MSE} = \frac{1}{n}\sum{(actual-forecast)^2}$
- RMSE, or Root Mean Squared Error
    - $\textrm{RMSE} = \sqrt{\frac{1}{n}\sum{(actual-forecast)^2}}$
- RMSE divided by average demand so it yeilds a % like MAPE
    - $\textrm{pRMSE} = \frac{\sqrt{\frac{1}{n}\sum{(actual-forecast)^2}}}{\frac{1}{n}\sum{actual}}$

It can be helpful to explicity caculate these to make comparison between datasets and models fair.  This section demonstration these calculation with SQL.

query = f"""
WITH
    FORECASTS AS (
        SELECT DATE({TIME_COLUMN}) as {TIME_COLUMN}, 
            DATE(predicted_on_date) as predicted_on_date, 
            CAST({TARGET_COLUMN} as INT64) AS {TARGET_COLUMN}, 
            #splits,
            {SERIES_COLUMN}, 
            predicted_{TARGET_COLUMN}.value as predicted_{TARGET_COLUMN}
        FROM `{PROJECT_ID}.{BQ_DATASET}.{BQ_TABLE}_automl`
    ),
    LEAD_DAYS AS (
        SELECT *, DATE_DIFF({TIME_COLUMN}, predicted_on_date, {FORECAST_GRANULARITY}) as prediction_lead_days
        FROM FORECASTS
    ),
    LATEST AS (
        SELECT {SERIES_COLUMN}, {TIME_COLUMN}, min(prediction_lead_days) as prediction_lead_days
        FROM LEAD_DAYS
        GROUP BY {SERIES_COLUMN}, {TIME_COLUMN}
    ),
    DIFFS AS (
        SELECT 
            {SERIES_COLUMN}, 
            {TIME_COLUMN}, 
            'forecast' as time_series_type,
            predicted_{TARGET_COLUMN} as forecast_value,
            {TARGET_COLUMN} as actual_value,
            ({TARGET_COLUMN} - predicted_{TARGET_COLUMN}) as diff
        FROM LATEST
        LEFT OUTER JOIN LEAD_DAYS
        USING ({SERIES_COLUMN}, {TIME_COLUMN}, prediction_lead_days)    
    )
SELECT {SERIES_COLUMN}, time_series_type, 
    AVG(SAFE_DIVIDE(ABS(diff), actual_value)) as MAPE,
    AVG(ABS(diff)) as MAE,
    SAFE_DIVIDE(SUM(ABS(diff)),SUM(actual_value)) as pMAE,
    AVG(POW(diff, 2)) as MSE,
    SQRT(AVG(POW(diff, 2))) as RMSE,
    SAFE_DIVIDE(SQRT( AVG( POW(diff, 2) ) ) , AVG(actual_value) ) as pRMSE
FROM DIFFS
GROUP BY {SERIES_COLUMN}, time_series_type
ORDER BY {SERIES_COLUMN}, time_series_type    
"""

# customMetrics = bq_client.query(query = query).to_dataframe()
# customMetrics

# print to run in BigQuery console
print(query)

Overall Metrics:

query = f"""
WITH
    FORECASTS AS (
        SELECT 
            DATE({TIME_COLUMN}) as {TIME_COLUMN}, 
            DATE(predicted_on_date) as predicted_on_date, 
            CAST({TARGET_COLUMN} as INT64) AS {TARGET_COLUMN}, 
            #splits, 
            {SERIES_COLUMN}, 
            predicted_{TARGET_COLUMN}.value as predicted_{TARGET_COLUMN}
        FROM `{PROJECT_ID}.{BQ_DATASET}.{BQ_TABLE}_automl`
    ),
    LEAD_DAYS AS (
        SELECT 
            *, 
            DATE_DIFF({TIME_COLUMN}, 
            predicted_on_date, 
            {FORECAST_GRANULARITY}) as prediction_lead_days
        FROM FORECASTS
    ),
    LATEST AS (
        SELECT 
            {SERIES_COLUMN}, 
            {TIME_COLUMN}, 
            min(prediction_lead_days) as prediction_lead_days
        FROM LEAD_DAYS
        GROUP BY {SERIES_COLUMN}, {TIME_COLUMN}
    ),
    DIFFS AS (
        SELECT 
            {SERIES_COLUMN}, 
            {TIME_COLUMN}, 
            'forecast' as time_series_type,
            predicted_{TARGET_COLUMN} as forecast_value,
            {TARGET_COLUMN} as actual_value,
            ({TARGET_COLUMN} - predicted_{TARGET_COLUMN}) as diff
        FROM LATEST
        LEFT OUTER JOIN LEAD_DAYS
        USING ({SERIES_COLUMN}, {TIME_COLUMN}, prediction_lead_days)    
    )
SELECT time_series_type, 
    AVG( SAFE_DIVIDE(ABS( diff ) , actual_value) ) as MAPE,
    AVG(ABS(diff)) as MAE,
    SAFE_DIVIDE(SUM( ABS( diff ) ) , SUM(actual_value) )  as pMAE,
    AVG(POW(diff, 2)) as MSE,
    SQRT(AVG(POW(diff, 2))) as RMSE,
    SAFE_DIVIDE(SQRT( AVG( POW(diff, 2) ) ) , AVG(actual_value) )  as pRMSE
FROM DIFFS
GROUP BY time_series_type
ORDER BY time_series_type    
"""
customMetricsOverall = bq_client.query(query = query).to_dataframe()
customMetricsOverall

# print to run in BigQuery console
# print(query)

# create run name
TIMESTAMP = datetime.now().strftime("%Y%m%d%H%M%S")
EXPERIMENT_RUN_NAME = f"run-{TIMESTAMP}"

# log params and metrics to dicts
params = {}
params["budget_hrs"] = MILLI_NODE_HRS
params["horizon"] = FORECAST_HORIZON
params["context_window"] = CONTEXT_WINDOW

metrics = {}
metrics["MAPE"] = customMetricsOverall['MAPE'][0]
metrics["MAE"] = customMetricsOverall['MAE'][0]
metrics["pMAE"] = customMetricsOverall['pMAE'][0]
metrics["MSE"] = customMetricsOverall['MSE'][0]
metrics["RMSE"] = customMetricsOverall['RMSE'][0]
metrics["pRMSE"] = customMetricsOverall['pRMSE'][0]

# # Create and log experiment
vertex_ai.init(experiment=EXPERIMENT_NAME.replace("_","-"))

with vertex_ai.start_run(EXPERIMENT_RUN_NAME) as my_run:
    my_run.log_metrics(metrics)
    my_run.log_params(params)

    vertex_ai.end_run()

## Get Forecasted Values for Future Horizon

## TODO
query = f"""
CREATE OR REPLACE TABLE `{PROJECT_ID}.{BQ_DATASET}.{BQ_TABLE_PLAN}_automl_batch_input` AS
WITH
    DATELIST AS (
        SELECT *
        FROM (SELECT DISTINCT {SERIES_COLUMN} FROM `{PROJECT_ID}.{BQ_DATASET}.{BQ_TABLE_PLAN}_source`) A
        CROSS JOIN (SELECT * 
                    FROM UNNEST(GENERATE_DATE_ARRAY(
                                    DATE_SUB((SELECT MAX({TIME_COLUMN}) FROM `{PROJECT_ID}.{BQ_DATASET}.{BQ_TABLE}_prepped`), INTERVAL {CONTEXT_WINDOW-1} {FORECAST_GRANULARITY}),
                                    DATE_ADD((SELECT MAX({TIME_COLUMN}) FROM `{PROJECT_ID}.{BQ_DATASET}.{BQ_TABLE}_prepped`), INTERVAL {FORECAST_HORIZON} {FORECAST_GRANULARITY}),
                                    INTERVAL 1 {FORECAST_GRANULARITY}
                                )
                            ) AS {TIME_COLUMN}
                    ) B
    ),
    ADDTARGET AS (
        SELECT *
        FROM DATELIST
        LEFT OUTER JOIN (SELECT {SERIES_COLUMN}, {TIME_COLUMN}, {TARGET_COLUMN} FROM `{PROJECT_ID}.{BQ_DATASET}.{BQ_TABLE_PLAN}_source`)
        USING ({SERIES_COLUMN}, {TIME_COLUMN})
        ORDER BY {SERIES_COLUMN}, {TIME_COLUMN}
    ),
    LOCF AS (
        SELECT 
          {SERIES_COLUMN}, 
          {TIME_COLUMN},
          LAST_VALUE({TARGET_COLUMN} IGNORE NULLS) OVER (PARTITION BY {SERIES_COLUMN} ORDER BY {TIME_COLUMN} ROWS BETWEEN UNBOUNDED PRECEDING AND CURRENT ROW) as {TARGET_COLUMN}
        FROM ADDTARGET
    )
SELECT 
    {SERIES_COLUMN}, 
    {TIME_COLUMN},
    CASE
        WHEN {TIME_COLUMN} > (SELECT MAX({TIME_COLUMN}) FROM `{PROJECT_ID}.{BQ_DATASET}.{BQ_TABLE_PLAN}_source`) THEN NULL
        ELSE {TARGET_COLUMN}
    END AS {TARGET_COLUMN}
FROM LOCF
ORDER BY {SERIES_COLUMN}, {TIME_COLUMN}
# """
job = bq_client.query(query = query)
job.result()

# print to run in BigQuery console
# print(query)

# Batch Prediction Job

# batchjob = vertex_ai.BatchPredictionJob.create(
batch_prediction_job = forecast.batch_predict(
    job_display_name = f'{EXPERIMENT_NAME}_automl_{CONTEXT_WINDOW}',
    # model_name = forecast.resource_name,
    instances_format = 'bigquery',
    predictions_format = 'bigquery',
    bigquery_source = f"bq://{PROJECT_ID}.{BQ_DATASET}.{BQ_TABLE_PLAN}_automl_batch_input",
    bigquery_destination_prefix = f"bq://{PROJECT_ID}.{BQ_DATASET}",
    sync=False
)

## Process Predicted Forecast

# batchjob.output_info.bigquery_output_table
batch_prediction_job.bigquery_output_table

batch_predict_bq_output_uri = "{}.{}".format(
    batch_prediction_job.output_info.bigquery_output_dataset,
    batch_prediction_job.output_info.bigquery_output_table
)

batch_predict_bq_output_uri

query = f"""
    CREATE OR REPLACE TABLE `{PROJECT_ID}.{BQ_DATASET}.{BQ_TABLE_PLAN}_automl_batch_output` AS
    SELECT {SERIES_COLUMN}, DATE({TIME_COLUMN}) as {TIME_COLUMN}, predicted_{TARGET_COLUMN}.value as predicted_{TARGET_COLUMN}
    FROM `{PROJECT_ID}.{BQ_DATASET}.{batch_prediction_job.output_info.bigquery_output_table}`
"""
job = bq_client.query(query = query)
job.result()

# print to run in BigQuery console
# print(query)

query = f"""
    SELECT *
    FROM `{PROJECT_ID}.{BQ_DATASET}.{BQ_TABLE_PLAN}_automl_batch_output`
    ORDER BY {SERIES_COLUMN}, {TIME_COLUMN}
"""
predict = bq_client.query(query = query).to_dataframe()
predict.head()

## Review Results

### Retrieve the Forecasting Data (raw series)


# TODO

# query = f"""
# SELECT start_station_name, date, splits, num_trips
# FROM `{PROJECT_ID}.{BQ_DATASET}.{BQ_TABLE_PLAN}_prepped`
# ORDER by start_station_name, date
# """
# rawSeries = bigquery.query(query = query).to_dataframe()
